requirements.txt
Flask
Werkzeug
Flask-SQLAlchemy
Flask-Login
SQLAlchemy
python-dotenv
psycopg2-binary
gunicorn
email-validator
bcrypt
